// src/app/model/ReportHeaderName.ts
export enum ReportHeaderName {
  id = 'Ид.номер',
  date = 'Дата',
  currencyInfo = 'Валюта',
// CrossProfit
  revenue = 'Выручка',
  costSales = 'Себестоимость',
  crossProfit = '',
// OperatingProfit
  sellingAndMarketingCost = '',
  administrativeExpenses = '',
  otherIncome = '',
  otherLosses = '',
  operatingProfit = '',
// NetProfit
  financeIncome = '',
  financeCosts = '',
  exchangeTransaction = '',
  incomeTax = '',
  ownersProfit = '',
  nonControlInterests = '',
  netProfit = '',
  netProfitBeforeTax = '',
// Other
  countStocks = '',
  profitPerStock = ''
}